requirements.txt
Django==4.2.2
django-cleanup==7.0.0
django-simple-captcha==0.5.17
django-widget-tweaks==1.4.12
Pillow==9.4.0

